requirements.txt
numpy==2.1.2
pandas==2.3.3
scikit-learn==1.7.2
matplotlib==3.10.7
matplotlib-inline==0.1.7
tqdm==4.67.1
pillow==11.0.0
segmentation_models_pytorch==0.5.0
seaborn==0.13.2